+++ src/index.ts
@@ -7,4 +7,3 @@ export * from "./lib/useFullyControlledReactDataTable/useFullyControlledReactDat
 export * from "./lib/utils/createReactDataTableColumnHelper";
 export * from "./lib/utils/getModelFromColumnFilter";
 export * from "./lib/utils/getColumnFilterFromModel";
-export * from "./lib/utils/getDraggableColumn";
